+++ benchmarks/verus/numpy_triple_part_2/files/mathematical_functions_clip.rs
@@ -8,20 +8,20 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn clip(arr: &Vec<i32>, min_val: i32, max_val: i32) -> (result: Vec<i32>)
+fn clip(arr: &Vec<i8>, min_val: i8, max_val: i8) -> (result: Vec<i8>)
     ensures
         result.len() == arr.len(),
         forall|i: int| 0 <= i < result.len() ==> {
-            if min_val <= max_val {
-                if arr[i] < min_val {
-                    result[i] == min_val
-                } else if arr[i] > max_val {
-                    result[i] == max_val
+            if (min_val as int) <= (max_val as int) {
+                if (arr[i] as int) < (min_val as int) {
+                    (result[i] as int) == (min_val as int)
+                } else if (arr[i] as int) > (max_val as int) {
+                    (result[i] as int) == (max_val as int)
                 } else {
-                    result[i] == arr[i]
+                    (result[i] as int) == (arr[i] as int)
                 }
             } else {
-                result[i] == max_val
+                (result[i] as int) == (max_val as int)
             }
         }
 // </vc-spec>
