+++ benchmarks/verus/numpy_triple_part_2/files/mathematical_functions_sinc.rs
@@ -8,17 +8,17 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn sinc(x: Vec<i32>) -> (result: Vec<i32>)
+fn sinc(x: Vec<i8>) -> (result: Vec<i8>)
     ensures 
         result.len() == x.len(),
         forall|i: int| 0 <= i < result.len() ==> {
             /* Boundedness: sinc values are bounded by [-1, 1] */
-            result[i] <= 1 &&
-            result[i] >= -1 &&
+            result[i] as int <= 1 &&
+            result[i] as int >= -1 &&
             /* Maximum at zero: sinc(0) = 1 */
-            (x[i] == 0 ==> result[i] == 1) &&
+            (x[i] as int == 0 ==> result[i] as int == 1) &&
             /* Symmetry: sinc is an even function */
-            (forall|j: int| 0 <= j < x.len() && x[i] == -x[j] ==> result[i] == result[j])
+            (forall|j: int| 0 <= j < x.len() && x[i] as int == -(x[j] as int) ==> result[i] as int == result[j] as int)
         }
 // </vc-spec>
 // <vc-code>
