+++ benchmarks/verus/numpy_triple_part_2/files/statistics_nanpercentile.rs
@@ -18,7 +18,7 @@ spec fn vec_sum(a: Seq<i32>) -> int
     }
 }
 
-fn nanpercentile(a: Vec<i32>, q: i32) -> (result: i32)
+fn nanpercentile(a: Vec<i8>, q: i8) -> (result: i8)
     requires 0 <= q <= 100,
     ensures
         /* Case 1: Empty array returns 0 */
@@ -26,7 +26,7 @@ fn nanpercentile(a: Vec<i32>, q: i32) -> (result: i32)
         /* Case 2: Non-empty array */
         a.len() > 0 ==> {
             /* Result is within bounds of input values */
-            (forall|i: int| 0 <= i < a.len() ==> a[i] <= result || result <= a[i]) &&
+            (forall|i: int| 0 <= i < a.len() ==> a[i] as int <= result as int || result as int <= a[i] as int) &&
             /* For single element, result equals that element */
             (a.len() == 1 ==> result == a[0])
         }
